% Generated by roxygen2: do not edit by hand
\docType{package}
\name{SNPmineR-package}
\alias{SNPmineR}
\title{SNPmineR: Tidyverse-Native Scaffold for SNP Orthology Mapping}
\description{Tidyverse-friendly API scaffold for mapping human SNP rsIDs to orthologous variants with explicit taxonomic scoping.}
