% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa_set.R
\name{taxa_set_from_species}
\alias{taxa_set_from_species}
\title{Create a TaxonSet from species codes}
\usage{
taxa_set_from_species(species_codes, name = "species")
}
\arguments{
\item{species_codes}{Character vector of species codes (e.g., "mmusculus").}

\item{name}{Optional name; defaults to "species".}
}
\value{
A \code{TaxonSet}.
}
\description{
Create a TaxonSet from species codes
}
