% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa_set.R
\name{taxa_set_from_taxids}
\alias{taxa_set_from_taxids}
\title{Create a TaxonSet from taxids}
\usage{
taxa_set_from_taxids(taxids, name = "taxids")
}
\arguments{
\item{taxids}{Integer vector of NCBI taxonomy IDs.}

\item{name}{Optional name; defaults to "taxids".}
}
\value{
A \code{TaxonSet}.
}
\description{
Create a TaxonSet from taxids
}
