% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_ortholog_db.R
\name{validate_ortholog_db}
\alias{validate_ortholog_db}
\title{Validate an ortholog mapping table}
\usage{
validate_ortholog_db(ortholog_db)
}
\arguments{
\item{ortholog_db}{A data.frame/tibble with required columns.}
}
\value{
Invisibly returns TRUE; errors if invalid.
}
\description{
Validate an ortholog mapping table
}
\keyword{internal}
