% Generated by roxygen2: do not edit by hand
\name{validate_variant_db}
\alias{validate_variant_db}
\alias{validate_ortholog_db}
\alias{validate_ref_db}
\title{Validate SNPmineR database schemas}
\usage{
validate_variant_db(variant_db)

validate_ortholog_db(ortholog_db)

validate_ref_db(refdb)
}
\value{Invisibly returns \code{TRUE} if valid.}
\description{Schema validators for variant, ortholog, and reference tables.}
