% Generated by roxygen2: do not edit by hand
\name{snp_rsids_to_consequence}
\alias{snp_rsids_to_consequence}
\title{Resolve rsIDs to human coding consequences}
\usage{
snp_rsids_to_consequence(
  rsids,
  variant_db,
  transcript_policy = c("canonical", "all", "pick_longest")
)
}
\value{A tibble with class \code{SNPmineHumanConsequence}.}
\description{Resolve rsIDs to human coding consequences.}
