% Generated by roxygen2: do not edit by hand
\name{taxa_set}
\alias{taxa_set}
\alias{taxa_set_from_taxids}
\alias{taxa_set_from_species}
\title{Create taxonomic scope objects}
\usage{
taxa_set(x)

taxa_set_from_taxids(taxids, name = "custom_taxids")

taxa_set_from_species(species_codes, name = "custom_species")
}
\value{A \code{TaxonSet} S3 object.}
\description{Construct taxonomic set objects for constrained querying.}
\examples{
taxa_set_from_taxids(c(40674, 7898), name = "Mammalia + Teleost")
taxa_set_from_species(c("mmusculus", "rnorvegicus"), name = "Rodents")
}
